requirements.txt
flask
slack_sdk
python-dotenv
langchain
langchain-community
langchain-openai
sentence-transformers
faiss-cpu
beautifulsoup4
requests
ollama
